+++ EHM.version
@@ -9,6 +9,6 @@
         "patch":4,
     },
     "starsectorVersion":"0.96a-RC10",
-    "directDownloadURL":"https://github.com/lyravega/lyr_ehm/releases/download/0.7.3/lyr_ehm.0.7.3.zip",
+    "directDownloadURL":"https://github.com/lyravega/lyr_ehm/releases/download/0.7.4/ExperimentalHullModifications.0.7.4.zip",
     "changelogURL":"https://raw.githubusercontent.com/lyravega/lyr_ehm/stable/changelog.txt"
 }
