+++ mod_info.json
@@ -5,6 +5,6 @@
 	"version":"0.7.4",
 	"description":"Adds experimental hull modifications",
 	"jars":["jars/lyr_ehm.jar"],
-	"modPlugin": "lyr.lyr_plugin",
+	"modPlugin": "lyravega.plugin.experimentalHullModifications",
 	"gameVersion":"0.96a-RC10"
 }
