+++ src/data/hullmods/ehm/events/normalEvents.java
@@ -8,7 +8,7 @@ import data.hullmods.ehm._ehm_basetracker;
  * When a change is detected through the {@link _ehm_basetracker tracker}, the event methods
  * in this interface will be fired if the hull modification's effect implements it.
  * <p> The hull modifications are registered as having this interface implemented during
- * {@link lyr.lyr_plugin#onApplicationLoad() onApplicationLoad()}
+ * {@link lyravega.plugin.experimentalHullModifications#onApplicationLoad() onApplicationLoad()}
  * @see {@link _ehm_basetracker Tracker Base} Provides tracking features for the fleet/ship
  * @author lyravega
  */
