+++ src/data/hullmods/ehm_ar/ehm_ar_stepdownadapter.java
@@ -17,8 +17,8 @@ import com.fs.starfarer.api.combat.WeaponAPI.WeaponSize;
 import com.fs.starfarer.api.ui.Alignment;
 import com.fs.starfarer.api.ui.TooltipMakerAPI;
 
-import lyr.misc.lyr_internals;
-import lyr.misc.lyr_tooltip;
+import lyravega.misc.lyr_internals;
+import lyravega.misc.lyr_tooltip;
 
 /**@category Adapter Retrofit 
  * @author lyravega
