+++ src/data/submarkets/ehm_submarket.java
@@ -13,14 +13,14 @@ import com.fs.starfarer.api.fleet.FleetMemberType;
 import com.fs.starfarer.api.impl.campaign.submarkets.BaseSubmarketPlugin;
 import com.fs.starfarer.api.loading.WeaponSpecAPI;
 
-import lyr.misc.lyr_internals;
+import lyravega.misc.lyr_internals;
 
 /**
  * A submarket for the experimental slot shunts. The submarket is attached/detached
  * whenever the player interacts with a valid market, and is not persistent.
  * <p> Each time it is attached to a market, it will get re-initialized, at which
  * point its contens will be refreshed.
- * @see {@link lyr.lyr_plugin.interactionListener interactionListener} for attaching/detaching the submarket
+ * @see {@link data.abilities.ehm_ability.ehm_interactionListener interactionListener} that reports when player opens/closes a market
  * @author lyravega
  */
 public class ehm_submarket extends BaseSubmarketPlugin {
