+++ src/lyravega/tools/_lyr_logger.java
@@ -1,8 +1,8 @@
-package lyr.tools;
+package lyravega.tools;
 
 import org.apache.log4j.Logger;
 
-import lyr.misc.lyr_internals;
+import lyravega.misc.lyr_internals;
 
 public interface _lyr_logger {
 	public static final Logger logger = Logger.getLogger(lyr_internals.logName);
