+++ src/lyravega/tools/_lyr_reflectionTools.java
@@ -1,4 +1,4 @@
-package lyr.tools;
+package lyravega.tools;
 
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
@@ -6,7 +6,7 @@ import java.lang.invoke.MethodHandles.Lookup;
 import java.lang.invoke.MethodType;
 import java.util.Arrays;
 
-import lyr.misc.lyr_internals;
+import lyravega.misc.lyr_internals;
 
 
 /**
